11/19

+++ app/src/main/java/com/example/justdoit/datas/ExpertReview.kt
@@ -1,4 +1,4 @@
 package com.example.justdoit.datas
 
-class ExpertReview(val nickname: String, val star: Int, val detail: String) {
+class ExpertReview(val nickname: String, val star: Float, val detail: String) {
 }
